chore: working on fixing overflow issues
bench:

+++ src/cli.rs
@@ -114,7 +114,7 @@ pub enum EngineCommand {
         square: Option<Square>,
 
         /// Evaluate `piece` at `square` with the provided endgame weight [0-100].
-        endgame_weight: Option<i32>,
+        endgame_weight: Option<u8>,
     },
 
     /// Performs a split perft on the current position at the supplied depth.
